% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_hip_fun.R
\name{apply_hip}
\alias{apply_hip}
\title{Aplicacao dos modelos hipsometricos}
\usage{
apply_hip(bd, coefs, priority = NULL, by.model_sel = F)
}
\arguments{
\item{bd}{dataframe com a base dados a ser aplicada os modelos. Output da funcao \code{\link{cal_var()}}}

\item{coefs}{lista dos coeficientes ajustados. Output da funcao \code{\link{fit_hip()}} ou \code{\link{coef_hip()}}}

\item{priority}{vetor com a prioridade de modelo a ser utilizado na altura predita. c("scolforo_simp", "scolforo", "pettersen", "curtis")}

\item{by.model_sel}{se TRUE sera utilizado o modelo pre-definido na na variavel  **model_sel** presente no dataframe **bd**. c("scolforo_simp", "scolforo", "pettersen", "curtis")}
}
\value{
Dataframe contendo as variaveis preditas com os modelos inputados
}
\description{
Aplicação dos modelos hipsometricos previamente ajustados \cr
}
\examples{

#OP1
priority <- c("scolforo_simp",
              "pettersen",
              "scolforo",
              "curtis")

bd <- apply_hip(bd, coefs = hip, priority = priority)

#OP2
bd$model_sel <- "scolforo"

bd <- apply_hip(bd, coefs = hip, priority = NULL, by.model_sel = T)

}
