% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_vol_fun.R
\name{apply_vol}
\alias{apply_vol}
\title{Caculos de volumes}
\usage{
apply_vol(
  bd,
  im = c("centro", "rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med"),
  cores = 4,
  h,
  dap,
  sortimento,
  h_toco,
  dwg,
  h_dwg,
  qbr,
  h_qbr,
  suprimir,
  b0,
  b1,
  b2,
  b3,
  b4,
  b5,
  b0_sc,
  b1_sc,
  b2_sc,
  b3_sc,
  b4_sc,
  b5_sc
)
}
\arguments{
\item{bd}{base de dados para aplicação do volume}

\item{im}{vetor com campos identificadores de medicao o padroa é \strong{Padrão será "centro", "rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med"}}

\item{cores}{n° de nucleos de processamento. \strong{Padrão sao 4}}

\item{h}{altura total da arvore - m}

\item{dap}{diametro a 1,3m - m}

\item{sortimento}{um data frame contendo a descrição dos sortimento. Para maiores informações checar a documentação da \code{\link{read_sort()}}.}

\item{h_toco}{altura de corte da arvore em metros. O Padrão é 0.1m}

\item{dwg}{se TRUE, sera considerado apenas o ultimo produto dos sortimentos ate a h_dwg. O padrão é FALSE}

\item{h_dwg}{altura em metros (caso dwg for TRUE) ate a qual sera realizado a extração de toras seguindo os produtos descritos no sortimento, apos essa altura sera apenas utilizado o ultimo sortimento (menor D.P.F.)}

\item{qbr}{se TRUE, sera simulada a extração de toras apenas ate a h_qbr. O padrao é falso FALSE.}

\item{h_qbr}{altura em metros (caso qbr for TRUE) ate a qual sera realizado a extração de toras. Caso qbr for TRUE e o valor de h_qbr for NA ou 0, esse sera calculado utilizando 70 da altura total}

\item{suprimir}{se TRUE, nao sera simulada a extração de nenhuma tora. Os valores de volume serao zerados. O padrão é FALSE. Aplica-se a DAPs e altura total zeradas, assim como falta de coeficientes da curva de afilamento}

\item{b0}{b0 do polinomio (Schöepfer, 1996)}

\item{b1}{b1 do polinomio (Schöepfer, 1996)}

\item{b2}{b2 do polinomio (Schöepfer, 1996)}

\item{b3}{b3 do polinomio (Schöepfer, 1996)}

\item{b4}{b4 do polinomio (Schöepfer, 1996)}

\item{b5}{b5 do polinomio (Schöepfer, 1996)}

\item{b0_sc}{b0 do polinomio sem casca (Schöepfer, 1996)}

\item{b1_sc}{b1 do polinomio sem casca (Schöepfer, 1996)}

\item{b2_sc}{b2 do polinomio sem casca (Schöepfer, 1996)}

\item{b3_sc}{b3 do polinomio sem casca (Schöepfer, 1996)}

\item{b4_sc}{b4 do polinomio sem casca (Schöepfer, 1996)}

\item{b5_sc}{b5 do polinomio sem casca (Schöepfer, 1996)}
}
\value{
um data frame contendo campos calculados com os volumes correspondentes
}
\description{
Realiza os caculos de volumes totais e comerciais abertos por sortimentos \cr
}
\examples{

apply_vol(bd,
          cores = 4,
          im = c("centro", "rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med"),
          h,
          dap,
          sortimento,
          h_toco,
          dwg,
          h_dwg,
          qbr,
          h_qbr,
          suprimir,
          b0,
          b1,
          b2,
          b3,
          b4,
          b5,
          b0_sc,
          b1_sc,
          b2_sc,
          b3_sc,
          b4_sc,
          b5_sc)

}
