% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by_log_fun.R
\name{by_log}
\alias{by_log}
\title{Gera a base final no nivel tora}
\usage{
by_log(
  bd,
  im = c("centro", "rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med"),
  ie = c(),
  cores = 4
)
}
\arguments{
\item{im}{vetor com campos identificadores de medicao.}

\item{cores}{n° de nucleos de processamento. \strong{Padrão sao 4}
\cr}

\item{db}{dataframe contendo as informacoes padronizadas do coletor - output da funcao \code{\link{poly2vol()}}}
}
\value{
um dataframe de mesma estrutura de \strong{bd}
}
\description{
Esta função ira organizar e gerar a base nivel tora \cr
}
\details{
\strong{Registros que nao possuirem correspondencia serao removidos da base de dados} \cr
}
\examples{

 db <- by_log(db, im, 4)

}
