% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cad_join_fun.R
\name{cad_join}
\alias{cad_join}
\title{Uni a base de dados do coletetor com o cadastro florestal}
\usage{
cad_join(dc, cad, by = c("rf", "talhao", "ciclo", "rotacao"), ...)
}
\arguments{
\item{dc}{dataframe contendo as informacoes padronizadas do coletor - output da funcao \code{\link{read_dc()}}}

\item{cad}{dataframe contendo as informacoes padronizadas do coletor - output da funcao \code{\link{read_cad()}}}

\item{by}{vetor de variaveis usadas para para unir as bases de dados}
}
\value{
um dataframe de mesma estrutura com novas variaveis unidas
}
\description{
Esta função unir a bade dados do coletor com a o cadastro florestal referente utilizando os campos descritos no parametro **by** \cr
}
\examples{

 dc <- cad_join(dc, cad)

}
