% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_var_fun.R
\name{cal_var}
\alias{cal_var}
\title{Calculo de variaveis}
\usage{
cal_var(x, by.assmann = FALSE, cor_area = FALSE)
}
\arguments{
\item{x}{dataframe contendo as informacoes padronizadas do coletor - output da funcao \code{\link{cad_join()}}}

\item{by.assmann}{o parametro by.assmann = FALSE calcula o hdom e ddom pelo metodo hibrido priorizando o codigo H quando presente cod1 e cod2. Caso nao tenha nenhum codigo H ele tenta realizar o calculo nas parcelas restantes utilizando o principio de Assmann(1970) a partir dos cap e alturas medidos. Quando este TRUE ocalculo ignora o codigo H e realiza o calculo apenas utilizando o principio de Assmann(1970)}

\item{cor_area}{quando o valor TRUE, indica que deve ser realizada a correção de area. E utilizado o valor maximo da inclinação para realização da correção. O campo **forma** deve conter a informacao `C` para indicar parcelas circulares, neste caso apenas a informacao do **lado1** é utiliada como raio da parcela.}
}
\value{
um dataframe de mesma estrutura com novas variaveis calculadas
}
\description{
Esta função ira calcular novas variaveis necessarias para ajuste e aplicação de hipsometrica \cr
}
\details{
Serão calculadas as variaveis a nivel de arvore:\cr
**dap** - diametro a 1,3m - cm\cr
**b** - area transversal - m²\cr
**h** - altura total - m\cr
\cr

 Serão calculadas as variaveis a nivel de populacao:\cr
**area_parc** - area da parcela - m²\cr
**dg** - diametro medio quadratico - cm\cr
**ab** - area basal - m²\cr
**hdom** - altura dominante - m\cr
**ddom** - diametro dominante - cm\cr
**idade** - idade - anos\cr

O agrupamento da populacao sera definido pelas variaveis c("rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med")\cr
}
\examples{

 dc <- cal_var(bd, by.assmann = F, cor_area = F)

}
