% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_hip_fun.R
\name{coef_hip}
\alias{coef_hip}
\title{Leitura dos coeficientes}
\usage{
coef_hip(folder_path)
}
\arguments{
\item{folder_path}{diretorio contendo os arquivos csvs dos coeficientes. Os arquivos csvs de input fazem parte do output da funcao \code{\link{fit_hip()}} e devem seguir o mesmo padrao de nomenclatura e variaveis.}
}
\value{
Um objeto tipo lista sera criado com os coeficientes presentes nos arquivos csvs
}
\description{
Esta função ira ler os coeficientes do arquivo csv e criar o objeto necessario para aplicação do modelo \cr
}
\examples{

hip <- coef_hip(folder_path)

}
