% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistency_fun.R
\name{consistency}
\alias{consistency}
\title{Relatorio de cosistencia}
\usage{
consistency(
  bd,
  path_outputs,
  im = c("centro", "rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med"),
  ie = c("especie", "matgen", "regime", "classe_idade"),
  it = c("centro", "rf", "talhao", "ciclo", "rotacao")
)
}
\arguments{
\item{bd}{dataframe contendo as informacoes padronizadas - output da funcao \code{\link{cal_var()}}}

\item{path_outputs}{diretorio onde sera escrito os outputs da funcao}

\item{im}{vetor com campos identificadores de medicao}

\item{ie}{vetor com campos identificadores de estrato}

\item{it}{vetor com campos identificadores de talhao}
}
\value{
sera emitido o relatorio em html no caminho indicado.
}
\description{
Esta função ira gerar o relatorio de consistencia \cr
}
\examples{

im <- c("centro", "rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med")
ie <- c("especie", "matgen", "regime", "classe_idade")
it <- c("centro", "rf", "talhao", "ciclo", "rotacao")

 consistency(bd, path_outputs, im, ie)

}
