% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hip_fun.R
\name{fit_hip}
\alias{fit_hip}
\title{Ajuste dos modelos}
\usage{
fit_hip(bd, proj_name, path_outputs, estrato)
}
\arguments{
\item{bd}{dataframe contendo as informacoes padronizadas - output da funcao \code{\link{cal_var()}}}

\item{proj_name}{objeto tipo texto com o nome do projeto}

\item{path_outputs}{diretorio onde sera escrito os outputs da funcao}

\item{estrato}{deve ser uma vetor contendo as variaveis que compoem os estratos dos dois grupos de modelos. O primeiro grupo se refere aos modelos de Scolforoe e o segundo grupo aos modelos de Pettersen e Curtis}
}
\value{
sera emitido o relatorio em html e arquivos de apoio no caminho indicado. Tambem sera retornado um objeto contendo os coeficientes ajustados utilizado para aplicacao dos modelos
}
\description{
Esta função ira ajustar os modelos hipsometricos \cr
}
\details{
Esta funcao ira ajustar 4 modelos, divididos em 2 grupos. Para cada grupo sera usado uma estratificacao especifica definida nos parametros\cr
É necessario o minimo de 8 observacoes para o ajuste.
}
\examples{

estrato <- list(c("especie", "matgen", "regime"),
                c("especie", "matgen", "regime", "classe_idade"))

 hip <- fit_hip(bd, proj_name = "fit", path_outputs, estrato)

}
