% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_poly_fun.R
\name{fit_poly}
\alias{fit_poly}
\title{Cubagem e afilamento}
\usage{
fit_poly(cub, path_outputs = NULL, dcom = NULL, at = NULL, qc = c(0, 500))
}
\arguments{
\item{cub}{dataframe contendo as informacoes padronizadas - output da funcao \code{\link{read_cuba()}}}

\item{path_outputs}{diretorio onde sera escrito os outputs da funcao}

\item{dcom}{diametro comercial na ponta fina em centimetros, quando NULL é mantido os valores da base original, quando definido o valor este é simulado por interpolação linear}

\item{at}{altura de toco em metros, quando NULL é mantido os valores da base original, quando definido o valor este é simulado por interpolação linear}

\item{qc}{limites classes de diametro para ajuste particionado da polinomial}
}
\value{
sera emitido o relatorio em html no caminho indicado.
}
\description{
Esta função ira realizar a cubagem dos dados e o ajustye da curva de afiliamento \cr
}
\examples{


 fit_poly(bd, path_outputs, dcom = 0, at = .1)

}
