% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance_log_fun.R
\name{glance_log}
\alias{glance_log}
\title{glance log}
\usage{
glance_log(x)
}
\arguments{
\item{x}{modelo (lm e nlm)}
}
\value{
um dataframe
}
\description{
Esta função ainda esta em construcao, por hora acredito que funcione para lm e nlm \cr
}
\examples{

glance2(mol)

}
