% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dispersion_fun.R
\name{plot_dispersion}
\alias{plot_dispersion}
\alias{plot_dispersion.lm}
\title{plot_dispersion}
\usage{
plot_dispersion(x, ...)

\method{plot_dispersion}{lm}(
  x,
  data = NULL,
  exp_variables = NULL,
  delog = F,
  ncol = NULL,
  nrow = NULL,
  title = NULL
)
}
\arguments{
\item{x}{modelo}

\item{data}{dataframe contendo as variaveis a serem plotadas em função do residuo}

\item{exp_variables}{vetor com as variavris que serao utilizadas para construção dos graficos}

\item{delog}{quando em TRUE, reverte a escala logaritimica das variaveis preditoras}
}
\value{
uma lista de graficos
}
\description{
Plotagem de dispersao
}
\details{
Esta funcao gera os graficos do residuo % do modelo em função de variaveis externas contidas no dataframe
}
\examples{

mol <- lm(mpg ~ hp + cyl, data = mtcars)
pd <- plot_dispersion(mol, data = mtcars, exp_variables = "hp")
pd

}
