% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_norm_fun.R
\name{plot_norm}
\alias{plot_norm}
\alias{plot_norm.lm}
\title{plot_norm}
\usage{
plot_norm(x, ...)

\method{plot_norm}{lm}(x, ncol = NULL, nrow = NULL, title = NULL)
}
\arguments{
\item{x}{modelo}
}
\value{
uma lista de graficos
}
\description{
Plotagem da normalidade
}
\details{
Esta funcao gera graficos de normalidade dos resisuos do seu modelo
}
\examples{

mol <- lm(mpg ~ hp + cyl, data = mtcars)
pn <- plot_norm(mol)
pn

}
