% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resid_fun.R
\name{plot_resid}
\alias{plot_resid}
\alias{plot_resid.lm}
\title{plot_resid}
\usage{
plot_resid(x, ...)

\method{plot_resid}{lm}(x, delog = F, ncol = NULL, nrow = NULL, title = NULL)
}
\arguments{
\item{x}{modelo}

\item{delog}{quando em TRUE, reverte a escala logaritimica das variaveis preditoras}
}
\value{
uma lista de graficos
}
\description{
Plotagem de residuos
}
\details{
Esta funcao gera 3 grupos de graficos: residuos %'s do modelo em funcao do ajustado, observado; do residuo padronizado em funcao do ajustado; observado em funcao do ajustado
}
\examples{

mol <- lm(mpg ~ hp + cyl, data = mtcars)
pr <- plot_resid(mol, delog = F)
pr

}
