% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly2log_fun.R
\name{poly2log}
\alias{poly2log}
\title{Simula a estração de toras de um fuste utilizando a equação de afilamento (Schöepfer, 1996)}
\usage{
poly2log(
  h,
  dap,
  sortimento,
  h_toco = 0.1,
  dwg = F,
  h_dwg,
  qbr = F,
  h_qbr,
  suprimir = F,
  b0 = b0,
  b1 = b1,
  b2 = b2,
  b3 = b3,
  b4 = b4,
  b5 = b5
)
}
\arguments{
\item{h}{altura total da arvore - m.}

\item{dap}{diametro a 1,3m - m.}

\item{sortimento}{um data frame contendo a descrição dos sortimento. Para maiores informações checar a documentação da \code{\link{read_sort()}}.}

\item{h_toco}{altura de corte da arvore em metros. O Padrão é 0.1m}

\item{dwg}{se TRUE, sera considerado apenas o ultimo produto dos sortimentos ate a h_dwg. O padrão é FALSE.}

\item{h_dwg}{altura em metros (caso dwg for TRUE) ate a qual sera realizado a extração de toras seguindo os produtos descritos no sortimento, apos essa altura sera apenas utilizado o ultimo sortimento (menor D.P.F.)}

\item{qbr}{se TRUE, sera simulada a extração de toras apenas ate a h_qbr. O padrao é falso FALSE.}

\item{h_qbr}{altura em metros (caso qbr for TRUE) ate a qual sera realizado a extração de toras. Caso qbr for TRUE e o valor de h_qbr for NA ou 0, esse sera calculado utilizando 70 da altura toral.}

\item{suprimir}{se TRUE, nao sera simulada a extração de nenhuma tora. Os valores de volume serao zerados. O padrão é FALSE. Aplica-se a DAPs e altura total zeradas, assim como falta de coeficientes da curva de afilamento.}

\item{b0}{b0 do polinomio (Schöepfer, 1996)}

\item{b1}{b1 do polinomio (Schöepfer, 1996)}

\item{b2}{b2 do polinomio (Schöepfer, 1996)}

\item{b3}{b3 do polinomio (Schöepfer, 1996)}

\item{b4}{b4 do polinomio (Schöepfer, 1996)}

\item{b5}{b5 do polinomio (Schöepfer, 1996)}
}
\value{
um data frame contendo a relação de toras extraidas na simulação
}
\description{
Essa função ira simular a estração de toras da arvore utilizando uma função de afilamento (Schöepfer, 1996) em funcao de um determinado diametro, altura total, altura de toco e descricao do sortimento apontada. \cr
}
\examples{


poly5_logs(dap, h, coef_poli, sortimento)

}
