% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_join_fun.R
\name{poly_join}
\alias{poly_join}
\title{Uni a base de dados com os coeficientes da função de afilamento}
\usage{
poly_join(db, coefs, ...)
}
\arguments{
\item{db}{dataframe contendo as informacoes padronizadas do coletor - output da funcao \code{\link{cal_var()}}}

\item{coefs}{dataframe contendo as informacoes padronizadas dos coeficientes da curva de afilamento - output da funcao \code{\link{read_poly()}}}
}
\value{
um dataframe de mesma estrutura com novas variaveis unidas
}
\description{
Esta função unir a bade dados com os coeficientes da função de afilamento utilizando os campos \strong{especie, matgen, regime e dap}
A condição utilizada é dap >= dap_1 & dap < dap_2
}
\details{
\strong{Registros que nao possuirem correspondencia serao removidos da base de dados} \cr
}
\examples{

 db <- poly_join(dc, coefs)

}
