% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_ifq_fun.R
\name{pp_ifq}
\alias{pp_ifq}
\title{Processamento IFQ}
\usage{
pp_ifq(
  x,
  cor_area = FALSE,
  im = c("rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med"),
  it = c("rf", "talhao", "ciclo", "rotacao")
)
}
\arguments{
\item{x}{dataframe contendo as informacoes padronizadas do coletor - output da funcao \code{\link{cad_join()}}}

\item{cor_area}{quando o valor TRUE, indica que deve ser realizada a correção de area. E utilizado o valor maximo da inclinação para realização da correção. O campo **forma** deve conter a informacao `C` para indicar parcelas circulares, neste caso apenas a informacao do **lado1** é utiliada como raio da parcela.}

\item{im}{vetor com campos identificadores de medicao.}

\item{it}{vetor com campos identificadores de talhao}
}
\value{
um dataframe de mesma estrutura com novas variaveis calculadas
}
\description{
Esta função ira processar o IFQ \cr
}
\examples{

 dc <- cal_var(bd,
               by.assmann = F,
               cor_area = F,
               im = c("rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med"),
               it = c("rf", "talhao", "ciclo", "rotacao"))

}
