% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cadifq_fun.R
\name{read_cadifq}
\alias{read_cadifq}
\title{Leitura do dados cadastrais para IFQ}
\usage{
read_cadifq(file, guess_max = 30000, ...)
}
\arguments{
\item{file}{caminho do csv2 contendo os dados cadastrais}

\item{guess_max}{estimativa do numero de observacoes da base de dados}
}
\value{
um dataframe similar ao csv2 de input com suas variaveis padroniadas
}
\description{
Esta função le os dados cadastrais utilizados para processamento de IFQ. Os dados devem ser salvos em um arquivo formatdo csv e ter como delimitador o \strong{";"} para para separacao decimal deve-se usar \strong{","} . \cr
Sao 26 campos obrigatorios e devem estar ordenados conforme a sequencia abaixo: \cr
\cr
\strong{centro} - indexador do centro \cr
\strong{rf} - identificacao de regiao florestal \cr
\strong{talhao} - identificacao do talhao \cr
\strong{ciclo} - identificacao do ciclo de plantio \cr
\strong{rotacao} - identificacao da rotacao de plantio \cr
\strong{especie} - descricao da especie \cr
\strong{matgen} - descricao do material genetico \cr
\strong{regime} - regime de manejo \cr
\strong{area_plt} - area plantada do talhao - ha \cr
\strong{dt_plt} - data de plantio - dd/mm/yyyy \cr
\strong{proj_silv} - cod do projeto de silvicultura \cr
\strong{solo} - tipo de solo \cr
\strong{pot_prod} - potencial produtivo \cr
\cr
Não importa a nomenclatura dos campos no arquivo csv, mas sim a sua posicao. \cr
O encording pre definido é o ISO-8859-1 - Latin alphabet \cr
campos adicionais podem ser adicionados apos a ultima coluna
}
\examples{

 dc <- read_cadifq("C:/Users/klem00125926/dados_coletor.csv")

}
