% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cuba_fun.R
\name{read_cuba}
\alias{read_cuba}
\title{Leitura de cubagem}
\usage{
read_cuba(file, guess_max = 1e+05, ...)
}
\arguments{
\item{file}{caminho do csv2 contendo os dados da cubagem}

\item{...}{argumentos da \link[=readr]{read_csv2}}
}
\value{
um dataframe similar ao csv2 de input com suas variaveis padroniadas
}
\description{
Esta função le os dados de campo. Os dados devem ser salvos em um arquivo formatdo csv e ter como delimitador o \strong{";"} e \strong{","} para separacao decimal. \cr
Sao 20 campos obrigatorios e devem estar ordenados conforme a sequencia abaixo: \cr
\cr
\strong{rf} - identificacao de regiao florestal \cr
\strong{talhao} - identificacao do talhao \cr
\strong{ciclo} - identificacao do ciclo de plantio \cr
\strong{rotacao} - identificacao da rotacao de plantio \cr
\strong{arvore} - identificacao da arvore \cr
\strong{dt_med} - data da medicao - dd/mm/yyyy \cr
\strong{idade} - idade da arvore -anos \cr
\strong{especie} - descricao da especie \cr
\strong{matgen} - descricao do material genetico \cr
\strong{regime} - regime de manejo \cr
\strong{toco} - altura de toco - m \cr
\strong{dap} - diametro a 1,3m - cm \cr
\strong{h} - altura total da arvore - m \cr
\strong{hi} - altura da seccao - m \cr
\strong{dcc1} - 1° diametro da seccao com casca - cm \cr
\strong{dcc2} - 2° diametro da seccao com casca - cm \cr
\strong{dsc1} - 1° diametro da seccao sem casca - cm \cr
\strong{dsc2} - 2° diametro da seccao sem casca - cm \cr
\strong{coleta} - realizador da coleta \cr
\strong{obs} - observacao \cr
\cr
Não importa a nomenclatura dos campos no arquivo csv, mas sim a sua posicao. \cr
O encording pre definido é o ISO-8859-1 - Latin alphabet \cr
}
\examples{

 dc <- read_cuba("dados_campo.csv")

}
