% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dc_fun.R
\name{read_dc}
\alias{read_dc}
\title{Leitura do dados de campo}
\usage{
read_dc(file, guess_max = 1e+05, ...)
}
\arguments{
\item{file}{caminho do csv2 contendo os dados de campo do coletor}

\item{...}{argumentos da readr::read_csv2}
}
\value{
um dataframe similar ao csv2 de input com suas variaveis padroniadas
}
\description{
Esta função le os dados de campo. Os dados devem ser salvos em um arquivo formatdo csv e ter como delimitador o **";"** e **","** para separacao decimal. \cr
Sao 26 campos obrigatorios e devem estar ordenados conforme a sequencia abaixo: \cr
\cr
\strong{atividade} - descricao da atividade \cr
\strong{rf} - identificacao de regiao florestal \cr
\strong{talhao} - identificacao do talhao \cr
\strong{ciclo} - identificacao do ciclo de plantio \cr
\strong{rotacao} - identificacao da rotacao de plantio \cr
\strong{dt_med} - data da medicao - dd/mm/yyyy \cr
\strong{lider} - nome do lider \cr
\strong{auxiliar} - nome do auxiliar \cr
\strong{parcela} - dentificacao da parcela \cr
\strong{tipo} - tipo de medicao \cr
\strong{forma} - forma da parcela \cr
\strong{hr_ini} - horario de inicio da medicao - HH:MM:SS \cr
\strong{hr_fim} - horario de finalizacao da medicao - HH:MM:SS \cr
\strong{coordX} - Coordenada geografica  \cr
\strong{coordY} - Coordenada geografica \cr
\strong{lado1} - Comprimento do lado da parcela -dm \cr
\strong{lado2} - Comprimento do lado da parcela -dm \cr
\strong{inc1} - Inclinacao da parcela - ° \cr
\strong{inc2} - Inclinacao da parcela - ° \cr
\strong{linha} - numero da linha \cr
\strong{arvore} - numero da arvore na linha \cr
\strong{cap} - circunferencia a 1,3m - mm \cr
\strong{alt} - altura total da arvore - dm \cr
\strong{cod1} - codigo qualitativo 1 \cr
\strong{cod2} - codigo qualitativo 2 \cr
\strong{codQ} - codigo qualitativo de tora \cr
\cr
Não importa a nomenclatura dos campos no arquivo csv, mas sim a sua posicao. \cr
O encording pre definido é o ISO-8859-1 - Latin alphabet \cr
}
\examples{

 dc <- read_dc("dados_coletor.csv")

}
