% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_poly_fun.R
\name{read_poly}
\alias{read_poly}
\title{Leitura dos coeficientes da curva de afilamento}
\usage{
read_poly(file, guess_max = 30000, ...)
}
\arguments{
\item{file}{caminho do csv2 contendo os dados cadastrais}

\item{guess_max}{estimativa do numero de observacoes da base de dados}
}
\value{
um dataframe similar ao csv2 de input com suas variaveis padroniadas
}
\description{
Esta função le os coeficientes do ajuste da curva de afilamento (Schöepfer, 1996). Os dados devem ser salvos em um arquivo formatdo csv e ter como delimitador o \strong{";"} para para separacao decimal deve-se usar \strong{","} . \cr
Sao 17 campos obrigatorios e devem estar ordenados conforme a sequencia abaixo: \cr
\cr
\strong{especie} - codificação de especie \cr
\strong{matgen} - codificacao de material genetico \cr
\strong{regime} - codificação do regime de manejo \cr
\strong{dap_min} - D.A.P. minimo - cm \cr
\strong{dap_max} - D.A.P. maximo - cm \cr
\strong{b0} - b0 do polinomio (Schöepfer, 1996) - ajuste com casca \cr
\strong{b1} - b1 do polinomio (Schöepfer, 1996) - ajuste com casca \cr
\strong{b2} - b2 do polinomio (Schöepfer, 1996) - ajuste com casca \cr
\strong{b3} - b3 do polinomio (Schöepfer, 1996) - ajuste com casca \cr
\strong{b4} - b4 do polinomio (Schöepfer, 1996) - ajuste com casca \cr
\strong{b5} - b5 do polinomio (Schöepfer, 1996) - ajuste com casca \cr
\strong{b0_sc} - b0 do polinomio (Schöepfer, 1996) - ajuste sem casca \cr
\strong{b1_sc} - b1 do polinomio (Schöepfer, 1996) - ajuste sem casca \cr
\strong{b2_sc} - b2 do polinomio (Schöepfer, 1996) - ajuste sem casca \cr
\strong{b3_sc} - b3 do polinomio (Schöepfer, 1996) - ajuste sem casca \cr
\strong{b4_sc} - b4 do polinomio (Schöepfer, 1996) - ajuste sem casca \cr
\strong{b5_sc} - b5 do polinomio (Schöepfer, 1996) - ajuste sem casca \cr
}
\details{
Não importa a nomenclatura dos campos no arquivo csv, mas sim a sua posicao. \cr
O encording pre definido é o ISO-8859-1 - Latin alphabet \cr
}
\examples{

 dc <- read_poly("C:/Users/klem00125926/descricao_sortimento.csv")

}
