% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sort_fun.R
\name{read_sort}
\alias{read_sort}
\title{Leitura dos padrões de sortimento}
\usage{
read_sort(file, guess_max = 30000, ...)
}
\arguments{
\item{file}{caminho do csv2 contendo os dados cadastrais}

\item{guess_max}{estimativa do numero de observacoes da base de dados}
}
\value{
um dataframe similar ao csv2 de input com suas variaveis padroniadas
}
\description{
Esta função le os padores de sortimento. Os dados devem ser salvos em um arquivo formatdo csv e ter como delimitador o \strong{";"}. Como separador decimal deve-se usar \strong{","} . \cr
}
\details{
Sao 7 campos obrigatorios e devem estar ordenados conforme descrito \cr
\strong{nome} - indexador do sortimento \cr
\strong{dpf} - diametro limite da ponta fina - cm. O valor sera arredondado para 2 casas decimais \cr
\strong{mincomp} - comprimento de tora minimo - m. O valor sera arredondado para 3 casas decimais \cr
\strong{maxcomp} - comprimento de tora maximo - m. O valor sera arredondado para 3 casas decimais. O menor comprimento possiver sera de 1 centimetro. \cr
\strong{padrao} - padrao de sortimento - esse campo chave deve corresponder ao valor atribuido na base de dados no campo "padrao" - campo numerico \cr
\strong{perda} - perda pela serra - cm \cr
\strong{molsap} - descricao do campo de atribuição na carga de modelo do SAP. Sendo as unicas classes disponiveis \strong{"c_vacel2", "c_vaclasse0", "c_vaclasse1", "c_vaclasse2", "c_vaclasse3", "p_vacel1", "p_vacel2", "p_vaclasse0", "p_vaclasse1"} \cr
\cr
Não importa a nomenclatura dos campos no arquivo csv, mas sim a sua posicao. \cr
O encording pre definido é o ISO-8859-1 - Latin alphabet \cr
}
\examples{

 dc <- read_sort("descricao_sortimento.csv")

}
