% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsm_fun.R
\name{rsm}
\alias{rsm}
\title{Gera a bases resumo}
\usage{
rsm(
  bd,
  grp = "estrato",
  sort,
  im = c("centro", "rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med"),
  ie = c("especie", "matgen", "regime", "classe_idade"),
  it = c("centro", "rf", "talhao", "ciclo", "rotacao")
)
}
\arguments{
\item{grp}{grupo de processameto. Os resumos do processamento pode ser realizado por estrato ou por talhao \strong{Padrão sera estrato}}

\item{sort}{dataframe contendo as informacoes padronizadas dos coeficientes da curva de afilamento - output da funcao \code{\link{read_sort()}}}

\item{im}{vetor com campos identificadores de medicao.\strong{Padrão sera c("centro", "rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med")}}

\item{ie}{vetor com campos identificadores de estrato.\strong{Padrão sera c("especie", "matgen", "regime", "classe_idade")}}

\item{it}{vetor com campos identificadores de talhao.\strong{Padrão sera c("centro", "rf", "talhao", "ciclo", "rotacao")}
\cr}

\item{db}{dataframe contendo as informacoes padronizadas do coletor - output da funcao \code{\link{poly2vol()}}}
}
\value{
uma lista de dataframe
}
\description{
Esta função ira organizar e gerar as bases para output \cr
}
\examples{

 db <- rsm(db, im, ie)

}
