% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_int_fun.R
\name{samp_int}
\alias{samp_int}
\title{Calculo da intensidade amostral}
\usage{
samp_int(
  x,
  cad = NULL,
  im = c("rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med"),
  ie = c("especie", "matgen", "regime", "classe_idade"),
  it = c("rf", "talhao", "ciclo", "rotacao")
)
}
\arguments{
\item{x}{dataframe contendo as informacoes padronizadas do coletor - output da funcao \code{\link{cad_join()}}}

\item{im}{vetor com campos identificadores de medicao}

\item{ie}{vetor com campos identificadores de estrato}

\item{it}{vetor com campos identificadores de talhao}
}
\value{
um dataframe com as informações de intensidade amostral de cada estrato
}
\description{
Esta função ira calcular a intensidade amostral \cr
}
\examples{

 ia <- samp_int(x = bd,
               im = c("rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med"),
               ie = c("especie", "matgen", "regime", "classe_idade"),
               it = c("rf", "talhao", "ciclo", "rotacao"))

}
