% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_join_fun.R
\name{sort_join}
\alias{sort_join}
\title{Uni a base de dados com os padroes de sortimento}
\usage{
sort_join(db, sort)
}
\arguments{
\item{db}{dataframe contendo as informacoes padronizadas do coletor - output da funcao \code{\link{cal_var()}} \strong{adicionado da variavel padrao}}

\item{sort}{dataframe contendo as informacoes padronizadas dos coeficientes da curva de afilamento - output da funcao \code{\link{read_sort()}}
\cr}
}
\value{
um dataframe de mesma estrutura com novas variaveis unidas
}
\description{
Esta função ira unir a bade dados com os padroes de sortimento pelo campo chave  \strong{padrao} \cr
}
\details{
\strong{Registros que nao possuirem correspondencia serao removidos da base de dados} \cr
}
\examples{

 db <- sort_join(db, sort)

}
