% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dc_fun.R
\name{write_dc}
\alias{write_dc}
\title{Escreve o csv de input do SAS}
\usage{
write_dc(bd, path_outputs)
}
\arguments{
\item{bd}{dataframe com a base dados a ser aplicada os modelos. Output da funcao \code{\link{apply_hip()}}}

\item{path_outputs}{caminho onde sera escrito o csv}
}
\value{
Escreve arquivo csv
}
\description{
Padroniza e escreve o csv formatado para input na rotina de processamento SAS \cr
}
\examples{

write_dc(bd, path_outputs)

}
