% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_int_fun.R
\name{samp_int}
\alias{samp_int}
\title{Calculo da intensidade amostral}
\usage{
samp_int(
  x,
  cad = NULL,
  var_est = c("especie", "matgen", "regime", "classe_idade")
)
}
\arguments{
\item{x}{dataframe contendo as informacoes padronizadas do coletor - output da funcao \code{\link{cad_join()}}}

\item{cad}{dataframe contendo as informacoes padronizadas do coletor \code{\link{read_cad()}}}

\item{var_est}{vetor contendo as variaveis que compoem o estrato}
}
\value{
um dataframe com as informações de intensidade amostral de cada estrato
}
\description{
Esta função ira calcular a intensidade amostral \cr
}
\details{
O agrupamento da populacao sera definido pelas variaveis c("rf", "talhao", "ciclo", "rotacao", "parcela", "dt_med")\cr
}
\examples{

 ia <- samp_int(x = bd, cad = NULL , var_est = c("especie", "matgen", "regime", "classe_idade"))

}
